upd project structure
